Add Minicircuits RC SP4T to public api
