+++ qcodes/instrument_drivers/Minicircuits/__init__.py
@@ -1,3 +1,4 @@
+from ._minicircuits_rc_sp4t import MiniCircuitsRCSP4T, MiniCircuitsRCSP4TChannel
 from .RUDAT_13G_90 import MiniCircuitsRudat13G90Usb
 from .USB_SPDT import MiniCircuitsSwitchChannelUsb, MiniCircuitsUsbSPDT
